Incorporates Alfonso's suggestions comments
