+++ app/src/main/java/uk/ac/ebi/atlas/utils/ExpressionAtlasBioentityIdentifiersReader.java
@@ -17,6 +17,7 @@ import uk.ac.ebi.atlas.model.experiment.differential.microarray.MicroarrayExperi
 import uk.ac.ebi.atlas.trader.ExperimentTrader;
 
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.util.HashSet;
 
 import static uk.ac.ebi.atlas.model.experiment.ExperimentType.MICROARRAY_1COLOUR_MICRORNA_DIFFERENTIAL;
@@ -136,11 +137,12 @@ public class ExpressionAtlasBioentityIdentifiersReader extends BioentityIdentifi
         return bioentityIdentifiers.size() - bioentityIdentifiersSizeWithoutNewElements;
     }
 
-    public HashSet<String> getBioentityIdsFromExperiment(String experimentAccession, boolean throwError) throws Exception {
+    public HashSet<String> getBioentityIdsFromExperiment(String experimentAccession, boolean throwError) {
         LOGGER.info("Reading gene IDs of {}", experimentAccession);
         Experiment experiment = experimentTrader.getExperimentForAnalyticsIndex(experimentAccession);
 
         HashSet<String> bioentityIdentifiers = new HashSet<>();
+        IOException e = null;
 
         if (experiment.getType().isBaseline()) {
 
@@ -151,11 +153,9 @@ public class ExpressionAtlasBioentityIdentifiersReader extends BioentityIdentifi
                     bioentityIdentifiers.add(analytics.geneId());
                     analytics = inputStream.readNext();
                 }
-            } catch (Exception exception) {
+            } catch (IOException exception) {
                 LOGGER.error(exception.getMessage());
-                if (throwError) {
-                    throw exception;
-                }
+                e = exception;
             }
 
         } else {  //if (experimentType.isDifferential()) {
@@ -173,9 +173,7 @@ public class ExpressionAtlasBioentityIdentifiersReader extends BioentityIdentifi
                         }
                     } catch (IOException exception) {
                         LOGGER.error(exception.getMessage());
-                        if (throwError) {
-                            throw exception;
-                        }
+                        e = exception;
                     }
                 }
 
@@ -190,12 +188,14 @@ public class ExpressionAtlasBioentityIdentifiersReader extends BioentityIdentifi
                     }
                 } catch (IOException exception) {
                     LOGGER.error(exception.getMessage());
-                    if (throwError) {
-                        throw exception;
-                    }
+                    e = exception;
                 }
             }
 
+        }
+
+        if (e != null && throwError) {
+            throw new UncheckedIOException(e);
         }
 
         return bioentityIdentifiers;
@@ -203,13 +203,6 @@ public class ExpressionAtlasBioentityIdentifiersReader extends BioentityIdentifi
 
     @Override
     public HashSet<String> getBioentityIdsFromExperiment(String experimentAccession) {
-        HashSet<String> bioentityIds = new HashSet<>();
-        try {
-            return getBioentityIdsFromExperiment(experimentAccession, false);
-        } catch (Exception e) {
-            // this will never be reached.
-            LOGGER.error("This should have not been reached:" + e.getMessage());
-        }
-        return bioentityIds;
+        return getBioentityIdsFromExperiment(experimentAccession, false);
     }
 }
