+++ cli/src/main/java/uk/ac/ebi/atlas/cli/experimentDesign/ExperimentDesignCommand.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Component;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
 
-import uk.ac.ebi.atlas.cli.utils.FailedAccessionWriter;
+import uk.ac.ebi.atlas.cli.utils.AccessionsWriter;
 import uk.ac.ebi.atlas.experimentimport.GxaExperimentCrud;
 
 import java.io.IOException;
